requirements.txt
matplotlib==3.5.1
numpy==1.21.2
Pillow==9.0.0
pandas==1.3.2
quadprog==0.1.11
scikit-learn==1.0.2
torch==1.11.0
torchvision==0.12.0
einops